Add wasm support, and support for 32 bit pointer abi

+++ src/libc/musl.c
@@ -29,7 +29,7 @@ SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 // Symbols expected by gcc/clang in freestanding mode.
 // Put them in separate sections to allow linker to discard unused ones.
 hc_SECTION(".text.memset")
-void *memset(void *dest, int32_t c, uint64_t n)
+void *memset(void *dest, int32_t c, size_t n)
 {
     unsigned char *s = dest;
     uint64_t k;
@@ -114,7 +114,7 @@ void *memset(void *dest, int32_t c, uint64_t n)
 
 
 hc_SECTION(".text.memmove")
-void *memmove(void *dest, const void *src, uint64_t n)
+void *memmove(void *dest, const void *src, size_t n)
 {
     typedef __attribute__((__may_alias__)) uint64_t WT;
 #define WS (sizeof(WT))
@@ -148,7 +148,7 @@ void *memmove(void *dest, const void *src, uint64_t n)
 }
 
 hc_SECTION(".text.memcpy")
-void *memcpy(void *restrict dest, const void *restrict src, uint64_t n)
+void *memcpy(void *restrict dest, const void *restrict src, size_t n)
 {
     unsigned char *d = dest;
     const unsigned char *s = src;
@@ -259,7 +259,7 @@ void *memcpy(void *restrict dest, const void *restrict src, uint64_t n)
 }
 
 hc_SECTION(".text.memcmp")
-int32_t memcmp(const void *vl, const void *vr, uint64_t n)
+int32_t memcmp(const void *vl, const void *vr, size_t n)
 {
     const unsigned char *l=vl, *r=vr;
     for (; n && *l == *r; n--, l++, r++);
